+++ dcag-api-service-java/src/main/java/com/ubr/dcagapiservicejava/parser/ImageLabellingTaskParser.java
@@ -17,6 +17,7 @@ import java.io.BufferedReader;
 import java.io.FileNotFoundException;
 import java.io.InputStreamReader;
 import java.io.Reader;
+import java.math.BigDecimal;
 import java.nio.channels.Channels;
 import java.util.HashSet;
 import java.util.Set;
@@ -30,7 +31,7 @@ public class ImageLabellingTaskParser implements TaskParser {
     String taskCurrency;
 
     @Value("${image_labelling_task_price}")
-    Double taskPrice;
+    BigDecimal taskPrice;
 
     @Value("${image_labelling_task_max_no_user}")
     Long maxNumberOfUser;
