+++ dcag-api-service-java/src/main/java/com/ubr/dcagapiservicejava/parser/RecordAudioTaskParser.java
@@ -17,6 +17,7 @@ import java.io.BufferedReader;
 import java.io.FileNotFoundException;
 import java.io.InputStreamReader;
 import java.io.Reader;
+import java.math.BigDecimal;
 import java.nio.channels.Channels;
 import java.util.HashSet;
 import java.util.Set;
@@ -29,7 +30,7 @@ public class RecordAudioTaskParser implements TaskParser {
     String taskCurrency;
 
     @Value("${record_audio_task_price}")
-    Double taskPrice;
+    BigDecimal taskPrice;
 
     @Value("${record_audio_task_max_no_user}")
     Long maxNumberOfUser;
